Combined InstructionUpload and Ad419Instruction controller.
