+++ AD419.DataHelper.Web/AD419.DataHelper.Web.csproj
@@ -230,7 +230,6 @@
     <Compile Include="Attributes\UseAntiForgeryTokenOnPostByDefault.cs" />
     <Compile Include="Attributes\ValidateAntiForgeryTokenHeaderOrFormAttribute.cs" />
     <Compile Include="Controllers\AccountController.cs" />
-    <Compile Include="Controllers\AD419InstructionsController.cs" />
     <Compile Include="Controllers\AllProjectsNewController.cs" />
     <Compile Include="Controllers\ArcCodeAccountExclusionsController.cs" />
     <Compile Include="Controllers\ArcCodesController.cs" />
@@ -243,7 +242,7 @@
     <Compile Include="Controllers\ExpenseOrgMappingsController.cs" />
     <Compile Include="Controllers\ExpiredProjectsController.cs" />
     <Compile Include="Controllers\FinalReportsController.cs" />
-    <Compile Include="Controllers\InstructionUploadController.cs" />
+    <Compile Include="Controllers\Ad419InstructionsController.cs" />
     <Compile Include="Controllers\PrincipalsController.cs" />
     <Compile Include="Controllers\ProjectMappingController.cs" />
     <Compile Include="Controllers\FieldStationExpensesImportController.cs" />
@@ -506,7 +505,7 @@
     <Content Include="Views\FinalReports\Index.cshtml" />
     <Content Include="Views\FinalReports\Report.cshtml" />
     <Content Include="Views\Shared\_LaborTransactionsForMissingCodes.cshtml" />
-    <Content Include="Views\InstructionUpload\Index.cshtml" />
+    <Content Include="Views\AD419Instructions\InstructionUpload.cshtml" />
     <Content Include="Views\AD419Instructions\Index.cshtml" />
   </ItemGroup>
   <ItemGroup>
