+++ AD419.DataHelper.Web/Views/AD419Instructions/Index.cshtml
@@ -1,9 +1,6 @@
-﻿
-@{
+﻿@{
     ViewBag.Title = "AD-419 Instructions";
 }
 
 <h2>@ViewBag.Title)</h2>
-
-
